Run #2944 sanitizer tests under Bun's native test runner
All new tests added for #2944 must execute under Bun's native test runner,
not Vitest. Move the sanitizeSequenceDiagram and sequenceDiagram prompt-
hardening cases into a dedicated *.bun.test.ts file that imports from
bun:test, register it in the Bun native test manifest so CI's Bun Native
Test Compatibility job runs it, and exclude *.bun.test.ts from the scripts
Vitest shard so the file isn't double-collected. The assertions themselves
are unchanged; only the runner changed.
